Moved cmocka tests to the right place, renamed test source files
