+++ ci/success.sh
@@ -2,7 +2,7 @@
 set -eu
 
 if [ "$BUILD_MODE" = "coverage" ]; then
-  cd src/cmocka
+  cd src/tests
   gcov-4.8 --object-file rnp_tests-rnp_tests.o rnp_tests.c
   bash <(curl -s https://codecov.io/bash)
 fi
