+++ remove_artifacts.sh
@@ -16,7 +16,7 @@ artifacts="
   m4/ltsugar.m4
   m4/ltversion.m4
   src/Makefile.in
-  src/cmocka/Makefile.in
+  src/tests/Makefile.in
   src/lib/Makefile.in
   src/lib/config.h.in
   src/lib/config.h.in~
