--- src/cmocka/password.txt
@@ -1,1 +1,0 @@
-maxpayne
