+++ src/tests/Makefile.am
@@ -3,8 +3,8 @@ bin_PROGRAMS		= rnp_tests
 rnp_tests_CPPFLAGS	= -I$(top_srcdir)/include -I$(top_srcdir)/src/lib $(JSONC_INCLUDES) $(BOTAN_INCLUDES)
 rnp_tests_LDADD		= ../lib/librnp.la ../rnpkeys/librnpkeys.la $(JSON_LIBS)
 rnp_tests_SOURCES   = \
-    rnp_tests_support.c \
-    rnp_tests_cipher.c \
-    rnp_tests_generatekey.c \
-    rnp_tests_exportkey.c \
+    support.c \
+    cipher.c \
+    generatekey.c \
+    exportkey.c \
     rnp_tests.c
